Fixes Replset Provider When Auth is Enabled
Previously, rs.init() would fail when auth is enabled because
the replset provider is connecting to the db using the --host
instead of localhost and forces auth.  This fails because
the user is yet to be created (as part of the mongodb::db class).

Users may not wish to require the db creation before the rs.init()