Adds Support for Authentication in Types and Providers
Previously, the types and providers did not function properly
when $auth was enabled.  Users would observe errors similar to:

"ok" : 0, "errmsg" : "unauthorized"

This is because the types/providers did not support authentication
credentials in the mongo shell commands.  This patch extends the
$auth parameter into the mongodb::db class and associated types/
providers to include the necessary credentials within the mongo
shell when $auth is true.